Converted client player rotation packet
